Enforce clustering enabled on node startup.
