+++ src/main/java/Launcher.java
@@ -1,4 +1,3 @@
-import Configuration.FileConfiguration;
 import Configuration.JsonFileStore;
 import Configuration.Strings;
 import Configuration.VertxSettings;
