+++ src/main/java/Logging/Server.java
@@ -2,36 +2,26 @@ package Logging;
 
 
 import Configuration.FileConfiguration;
-import Configuration.VertxSettings;
 import Logging.Configuration.LogProvider;
 import Logging.Configuration.LogServerSettings;
 import Logging.Controller.LogHandler;
 import Logging.Model.DefaultLogger;
-import Logging.Model.Logger;
 import Protocols.ClusterListener;
-import Protocols.ClusterServer;
+import Protocols.ClusterVerticle;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
-import io.vertx.core.Verticle;
 import io.vertx.core.Vertx;
 
 /**
  * @author Robin Duda
  *         Receives logging data from the other components and writes it to an elasticsearch cluster or console.
  */
-public class Server implements Verticle {
-    private Logger logger;
-    private Vertx vertx;
+public class Server extends ClusterVerticle {
     private LogServerSettings settings;
 
     @Override
-    public Vertx getVertx() {
-        return vertx;
-    }
-
-    @Override
     public void init(Vertx vertx, Context context) {
-        this.vertx = vertx;
+        super.init(vertx, context);
         this.settings = FileConfiguration.instance().getLogSettings();
         this.logger = new DefaultLogger(vertx, settings.getLogserver());
     }
@@ -45,11 +35,5 @@ public class Server implements Verticle {
         }
 
         logger.onServerStarted(start);
-    }
-
-
-    @Override
-    public void stop(Future<Void> stop) throws Exception {
-        logger.onServerStopped(stop);
     }
 }
